protect functions, norm correction

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: cde-neef <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/04 16:49:26 by cde-neef          #+#    #+#             */
-/*   Updated: 2016/11/04 16:53:29 by cde-neef         ###   ########.fr       */
+/*   Updated: 2016/11/12 18:01:14 by cde-neef         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t		ft_strlen(const char *s)
 {
 	const char *buffer;
